+++ source/R-Portable-Mac/library/openssl/tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(openssl)
+
+test_check("openssl")